♻️ Meilleure découpe controller legacy getProjectPage
